+++ packages/ui/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./lib/utils";
+export * from "./icons";
 export * from "./components/button";
 export * from "./components/card";
 export * from "./components/badge";